Give information about cells revealed by scrolls of magic mapping
After a scroll is used, the next call to getworld on that level will
contain tile info for all mapped cells. Subsequent calls will
be as normal. lastseen is set to -1 in the world table to indicate that
the tile was never seen. Also, only dungeon and liquid layers are given,
as per Brogue internals.